MatchPoint replaces MatchesHelper.
With this change the notion of 30-day and 60-day rankings goes away in
favor of a single all-time ranking that uses the official ratings
process used by the National Association of Table Tennis. The new
ranking is put next to the name of the person in the rankings for easy
reference.

This also removes a few defects associated with PivotPong, including
rankings getting messed up on removal of old matches as rank isn't
persisted.